Mobile_App_Churn_Prediction_&_Retention/requirements.txt
kafka-python
pyodbc
pandas
scikit-learn
xgboost
sqlalchemy
fastapi
uvicorn
streamlit
prometheus_client
plotly
